Footer code cleanup
