+++ blocks/footer/footer.css
@@ -140,10 +140,6 @@
   margin: 0;
 }
 
-.footer-item .footer-link-hidden {
-  display: none;
-}
-
 .footer-container {
   margin: 0 auto;
 }
