+++ blocks/footer/footer.js
@@ -1,32 +1,26 @@
-import { isSignedInUser } from '../../scripts/auth/profile.js';
 import { decorateIcons } from '../../scripts/lib-franklin.js';
-import { getPathDetails, fetchGlobalFragment, decorateLinks, decorateNewTabLinks } from '../../scripts/scripts.js';
+import { getPathDetails, fetchGlobalFragment, htmlToElement } from '../../scripts/scripts.js';
 import LanguageBlock from '../language/language.js';
 
-/**
- * Adds attributes to <a> tags based on special keys in the URL.
- *
- * If a URL contains '@auth-only', it adds auth-only="true".
- * Example:
- * <a href="https://example.com@auth-only"> → <a href="https://example.com" auth-only="true">
- *
- * @param {HTMLElement} block
- */
-const decorateFooterAuthOnlyLinks = (block) => {
-  block.querySelectorAll('a[href*="@auth-only"]').forEach((link) => {
-    link.href = link.href.replace('@auth-only', '');
-    link.setAttribute('auth-only', 'true');
+/** @param {HTMLElement} block  */
+export function decorateFooterLinks(block) {
+  const links = block.querySelectorAll('a[href*="@newtab"]');
+  links.forEach((link) => {
+    link.href = link.href.replace('@newtab', '');
+    link.setAttribute('target', '_blank');
+    link.setAttribute('rel', 'noopener noreferrer');
+    // insert before first text child node
+    const icon = htmlToElement('<span class="icon icon-link-out"></span>');
+    link.firstChild.after(icon);
+    decorateIcons(link);
   });
-};
+}
 
 async function decorateMenu(footer) {
-  const isSignedIn = await isSignedInUser();
   const childElements = footer.querySelectorAll('.footer-item');
   const groupDiv = document.createElement('div');
   groupDiv.classList.add('footer-menu');
-  decorateFooterAuthOnlyLinks(footer);
-  decorateLinks(footer);
-  decorateNewTabLinks(footer);
+  decorateFooterLinks(footer);
   childElements.forEach((child) => {
     const h2Elements = Array.from(child.querySelectorAll('h2'));
     const ulElements = Array.from(child.querySelectorAll('ul'));
@@ -37,23 +31,6 @@ async function decorateMenu(footer) {
         divPair.setAttribute('class', 'footer-item-list');
         divPair.appendChild(h2Element);
         const ulElement = ulElements[index];
-        const anchorLinks = Array.from(ulElement.querySelectorAll('a') ?? []);
-        const containsAuthOnlyLink = !!anchorLinks.find((a) => a.getAttribute('auth-only'));
-        if (containsAuthOnlyLink) {
-          const loginLink = anchorLinks.find((a) => a.getAttribute('auth-only') !== 'true');
-          if (loginLink) {
-            loginLink.href = '';
-            loginLink.classList.add('footer-login-link');
-          }
-          anchorLinks.forEach((a) => {
-            const authOnlyAttribute = a.getAttribute('auth-only');
-            const isSignedInAndUnAuthenticatedLink = isSignedIn && authOnlyAttribute !== 'true';
-            const isNotSignedInAndAuthenticatedLink = !isSignedIn && authOnlyAttribute === 'true';
-            if (isSignedInAndUnAuthenticatedLink || isNotSignedInAndAuthenticatedLink) {
-              a.classList.add('footer-link-hidden');
-            }
-          });
-        }
         divPair.appendChild(ulElement);
         divWrapper.appendChild(divPair);
       });
@@ -201,14 +178,6 @@ function handleSocialIconStyles(footer) {
   });
 }
 
-function handleLoginFunctionality(footer) {
-  const loginLink = footer.querySelector('.footer-login-link');
-  loginLink.addEventListener('click', (e) => {
-    e.preventDefault();
-    window.adobeIMS.signIn();
-  });
-}
-
 /**
  * loads and decorates the footer
  * @param {Element} block The footer block element
@@ -230,6 +199,5 @@ export default async function decorate(block) {
     await decorateMenu(footer);
     decorateCopyrightsMenu(footer);
     handleSocialIconStyles(footer);
-    handleLoginFunctionality(footer);
   }
 }
