+++ blocks/header/header.js
@@ -7,7 +7,6 @@
 import {
   htmlToElement,
   decorateLinks,
-  decorateNewTabLinks,
   getConfig,
   getLink,
   getPathDetails,
@@ -634,6 +633,18 @@ const adobeLogoDecorator = async (adobeLogoBlock) => {
   decorateIcons(adobeLogoBlock);
   return adobeLogoBlock;
 };
+
+/** @param {HTMLElement} block  */
+export function decorateNewTabLinks(block) {
+  const links = block.querySelectorAll('a[target="_blank"]');
+  links.forEach((link) => {
+    link.setAttribute('rel', 'noopener noreferrer');
+    // insert before first text child node
+    const icon = htmlToElement('<span class="icon icon-link-out"></span>');
+    link.firstChild.after(icon);
+    decorateIcons(link);
+  });
+}
 
 /**
  * Main header decorator, calls all the other decorators
